main file renamed, classes moved to separate files
